Updated answers, companies, questions, tags controllers

+++ app/controllers/admins/answers_controller.rb
@@ -3,9 +3,10 @@
 module Admins
   class AnswersController < AdminsController
     before_action :question
+
     def create
       @answer = @question.answers.build(answer_params)
-
+      authorize @answer
       if @answer.save
         AnswerMailer.with(answer: @answer, question: @question, admin: current_user).question_answered.deliver_later
         redirect_to admins_question_path(@question), notice: 'Answer created!'
@@ -17,6 +18,7 @@ module Admins
 
     def destroy
       answer = @question.answers.find(params[:id])
+      authorize @answer
       if answer.destroy
         redirect_to admins_question_path, notice: 'Answer destroyed!'
       else
